Implement keep_old_files option

+++ lib/mini_paperclip/storage/filesystem.rb
@@ -27,6 +27,7 @@ module MiniPaperclip
       end
 
       def do_delete_files
+        return if @deletes.empty?
         debug("deleting by filesystem #{@deletes}")
         FileUtils.rm_f(@deletes)
       end
